+++ tests/testthat/test-write.R
@@ -298,8 +298,7 @@ test_that("xportr_write Test 16: `split_by` attribute is used to split the data"
   expect_warning(
     dts |>
       xportr_split(split_by = "X") |>
-      xportr_write(path = tmp),
-    class = "lifecycle_warning_deprecated"
+      xportr_write(path = tmp)
   )
 
   expect_true(